Media type rearranged. UserByUsername type
