+++ models/media.ts
@@ -1,0 +1,7 @@
+export type TMediaType = {
+  id: string;
+  identity_id: string;
+  filename: string;
+  url: string;
+  created_at: string;
+};
